Refuse to trade with China
I'd like to just make the process limited in I/O as if there were
tariff's and other things of that nature imposed. But according to a
comment on this S.O answer [1] it wouldn't be cross platform and I don't
want to introduce a dependency on psutil.

If China is detected as the locale for the computer, we refuse to run
the program. After all, we can't let them keep beating us!

[1]:http://stackoverflow.com/a/6245160/1808164